# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"PlayerStatus": {
				"type": "string",
				"enum": [
					"draft",
					"team",
					"opponent",
					"scout",
					"transfer",
					"sold",
					"rejected"
				]
			},
			"TransferListedPlayer": {
				"properties": {
					"id": {
						"type": "string"
					},
					"game_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/PlayerStatus"
					},
					"throwing": {
						"type": "number",
						"format": "double"
					},
					"catching": {
						"type": "number",
						"format": "double"
					},
					"dodging": {
						"type": "number",
						"format": "double"
					},
					"blocking": {
						"type": "number",
						"format": "double"
					},
					"speed": {
						"type": "number",
						"format": "double"
					},
					"positional_sense": {
						"type": "number",
						"format": "double"
					},
					"teamwork": {
						"type": "number",
						"format": "double"
					},
					"clutch_factor": {
						"type": "number",
						"format": "double"
					},
					"throwing_potential": {
						"type": "number",
						"format": "double"
					},
					"catching_potential": {
						"type": "number",
						"format": "double"
					},
					"dodging_potential": {
						"type": "number",
						"format": "double"
					},
					"blocking_potential": {
						"type": "number",
						"format": "double"
					},
					"speed_potential": {
						"type": "number",
						"format": "double"
					},
					"positional_sense_potential": {
						"type": "number",
						"format": "double"
					},
					"teamwork_potential": {
						"type": "number",
						"format": "double"
					},
					"clutch_factor_potential": {
						"type": "number",
						"format": "double"
					},
					"tier": {
						"type": "number",
						"format": "double"
					},
					"potential_tier": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"buy_price": {
						"type": "number",
						"format": "double"
					},
					"sell_price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"game_id",
					"name",
					"status",
					"throwing",
					"catching",
					"dodging",
					"blocking",
					"speed",
					"positional_sense",
					"teamwork",
					"clutch_factor",
					"throwing_potential",
					"catching_potential",
					"dodging_potential",
					"blocking_potential",
					"speed_potential",
					"positional_sense_potential",
					"teamwork_potential",
					"clutch_factor_potential",
					"tier",
					"potential_tier",
					"created_at",
					"updated_at",
					"buy_price",
					"sell_price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetTransferListResponse": {
				"description": "Response model for retrieving transfer list",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"message": {
						"type": "string",
						"description": "Descriptive message about the operation result"
					},
					"transfer_list": {
						"items": {
							"$ref": "#/components/schemas/TransferListedPlayer"
						},
						"type": "array",
						"description": "List of players available for transfer"
					},
					"season": {
						"type": "number",
						"format": "double",
						"description": "Current season number"
					}
				},
				"required": [
					"success",
					"message",
					"transfer_list",
					"season"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Player": {
				"properties": {
					"id": {
						"type": "string"
					},
					"game_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/PlayerStatus"
					},
					"throwing": {
						"type": "number",
						"format": "double"
					},
					"catching": {
						"type": "number",
						"format": "double"
					},
					"dodging": {
						"type": "number",
						"format": "double"
					},
					"blocking": {
						"type": "number",
						"format": "double"
					},
					"speed": {
						"type": "number",
						"format": "double"
					},
					"positional_sense": {
						"type": "number",
						"format": "double"
					},
					"teamwork": {
						"type": "number",
						"format": "double"
					},
					"clutch_factor": {
						"type": "number",
						"format": "double"
					},
					"throwing_potential": {
						"type": "number",
						"format": "double"
					},
					"catching_potential": {
						"type": "number",
						"format": "double"
					},
					"dodging_potential": {
						"type": "number",
						"format": "double"
					},
					"blocking_potential": {
						"type": "number",
						"format": "double"
					},
					"speed_potential": {
						"type": "number",
						"format": "double"
					},
					"positional_sense_potential": {
						"type": "number",
						"format": "double"
					},
					"teamwork_potential": {
						"type": "number",
						"format": "double"
					},
					"clutch_factor_potential": {
						"type": "number",
						"format": "double"
					},
					"tier": {
						"type": "number",
						"format": "double"
					},
					"potential_tier": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"game_id",
					"name",
					"status",
					"throwing",
					"catching",
					"dodging",
					"blocking",
					"speed",
					"positional_sense",
					"teamwork",
					"clutch_factor",
					"throwing_potential",
					"catching_potential",
					"dodging_potential",
					"blocking_potential",
					"speed_potential",
					"positional_sense_potential",
					"teamwork_potential",
					"clutch_factor_potential",
					"tier",
					"potential_tier",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BuyTransferListedPlayerResponse": {
				"description": "Response model for buying a transfer-listed player",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"message": {
						"type": "string",
						"description": "Descriptive message about the operation result"
					},
					"player": {
						"$ref": "#/components/schemas/Player",
						"description": "Details of the purchased player"
					},
					"budget_remaining": {
						"type": "number",
						"format": "double",
						"description": "Remaining budget after the purchase"
					}
				},
				"required": [
					"success",
					"message",
					"player",
					"budget_remaining"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BuyTransferListedPlayerRequest": {
				"description": "Request model for buying a transfer-listed player",
				"properties": {
					"player_id": {
						"type": "string",
						"description": "ID of the player to buy"
					}
				},
				"required": [
					"player_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SellPlayerResponse": {
				"description": "Response model for selling a player",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Indicates if the operation was successful"
					},
					"message": {
						"type": "string",
						"description": "Descriptive message about the operation result"
					},
					"sold_player": {
						"$ref": "#/components/schemas/Player",
						"description": "Details of the sold player"
					},
					"budget": {
						"type": "number",
						"format": "double",
						"description": "Updated team budget after selling the player"
					}
				},
				"required": [
					"success",
					"message",
					"sold_player",
					"budget"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SellPlayerRequest": {
				"description": "Request model for selling a player",
				"properties": {
					"player_id": {
						"type": "string",
						"description": "ID of the player to sell"
					}
				},
				"required": [
					"player_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GameStatus": {
				"type": "string",
				"enum": [
					"pending",
					"in_progress",
					"completed"
				]
			},
			"GameStage": {
				"type": "string",
				"enum": [
					"draft",
					"pre_season",
					"regular_season",
					"post_season",
					"off_season"
				]
			},
...
			"CurrentGameResponseModel": {
				"description": "Current game response model for tsoa documentation",
				"properties": {
					"game_id": {
						"type": "string",
						"description": "Game ID",
						"example": "550e8400-e29b-41d4-a716-446655440000"
					},
					"game_season": {
						"type": "number",
						"format": "double",
						"description": "Game season number",
						"example": 1
					},
					"game_match_day": {
						"type": "number",
						"format": "double",
						"description": "Game match day",
						"example": 3
					},
					"game_status": {
						"$ref": "#/components/schemas/GameStatus",
						"description": "Game status",
						"example": "in_progress"
					},
					"game_stage": {
						"$ref": "#/components/schemas/GameStage",
						"description": "Game stage",
						"example": "regular_season"
					},
					"team_id": {
						"type": "string",
						"description": "Team ID",
						"example": "550e8400-e29b-41d4-a716-446655440001"
					},
					"team_name": {
						"type": "string",
						"description": "Team name",
						"example": "Dodgeball Dynamos"
					},
					"success": {
						"type": "boolean",
						"description": "Success status",
						"example": true
					},
					"message": {
						"type": "string",
						"description": "Response message",
						"example": "Current game retrieved successfully"
					}
				},
				"required": [
					"game_id",
					"game_season",
					"game_match_day",
					"game_status",
					"game_stage",
					"team_id",
					"team_name",
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "dodgeball-manager-api",
		"version": "1.0.0",
		"description": "Dodgeball manager game API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/transfers": {
			"get": {
				"operationId": "GetTransferList",
				"responses": {
					"200": {
						"description": "Transfer list retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetTransferListResponse"
								}
							}
						}
					}
				},
				"description": "Get the list of players available for transfer",
				"tags": [
					"Transfers"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/transfers/buy": {
			"post": {
				"operationId": "BuyPlayer",
				"responses": {
					"200": {
						"description": "Player bought successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BuyTransferListedPlayerResponse"
								}
							}
						}
					}
				},
				"description": "Buy a player from the transfer list",
				"tags": [
					"Transfers"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BuyTransferListedPlayerRequest"
							}
						}
					}
				}
			}
		},
		"/transfers/sell": {
			"post": {
				"operationId": "SellPlayer",
				"responses": {
					"200": {
						"description": "Player sold successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SellPlayerResponse"
								}
							}
						}
					}
				},
				"description": "Sell a player from the team",
				"tags": [
					"Transfers"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SellPlayerRequest"
							}
						}
					}
				}
			}
		},
...
		"/games/current": {
			"get": {
				"operationId": "GetCurrentGame",
				"responses": {
					"200": {
						"description": "Current game retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CurrentGameResponseModel"
								}
							}
						}
					}
				},
				"description": "Get the current game for the authenticated user",
				"summary": "Get current game",
				"tags": [
					"Games"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}